fix: support biginteger

+++ MiniGamesBox Classic/src/main/java/plugily/projects/minigamesbox/classic/commands/arguments/game/StatsArgument.java
@@ -62,7 +62,7 @@ public class StatsArgument {
             continue;
           }
           String statisticKey = statisticType.getName().toUpperCase();
-          new MessageBuilder("LEADERBOARD_STATISTICS_" + statisticKey).asKey().player(user.getPlayer()).integer((int)(user.getStatistic(statisticType))).send(sender);
+          new MessageBuilder("LEADERBOARD_STATISTICS_" + statisticKey).asKey().player(user.getPlayer()).integer(user.getStatistic(statisticType)).send(sender);
         }
       }
